+++ src/main/neuralnet/activations/TanH.java
@@ -7,24 +7,24 @@ public class TanH implements Activation {
 		return ActivationType.TANH;
 	}
 
-	public void activation(double[] x) {
-		IntStream.range(0, x.length).parallel().forEach(i -> x[i] = Math.tanh(x[i]));
+	public void activation(float[] x) {
+		IntStream.range(0, x.length).parallel().forEach(i -> x[i] = (float) Math.tanh(x[i]));
 	}
 
-	public void activation(double[][] x) {
+	public void activation(float[][] x) {
 		IntStream.range(0, x.length).parallel().forEach(b -> {
 			for (int i = 0; i < x[0].length; i++) {
-				x[b][i] = Math.tanh(x[b][i]);
+				x[b][i] = (float) Math.tanh(x[b][i]);
 			}
 		});
 	}
 
-	public double[][] derivative(double[][] x) {
-		double[][] derivative = new double[x.length][x[0].length];
+	public float[][] derivative(float[][] x) {
+		float[][] derivative = new float[x.length][x[0].length];
 
 		IntStream.range(0, x.length).parallel().forEach(b -> {
 			for (int i = 0; i < x[0].length; i++) {
-				derivative[b][i] = 1 - Math.pow(x[b][i], 2);
+				derivative[b][i] = 1 - (float) Math.pow(x[b][i], 2);
 			}
 		});
