+++ src/main/neuralnet/costs/CrossEntropy.java
@@ -13,16 +13,16 @@ public class CrossEntropy implements Cost {
 		return CostType.CROSS_ENTROPY;
 	}
 
-	public double cost(double[] out, double[] target) {
-		double cost = IntStream.range(0, target.length).parallel().mapToDouble(i -> (target[i] * Math.log(out[i] + 1e-16))).sum();
+	public float cost(float[] out, float[] target) {
+		float cost = (float) IntStream.range(0, target.length).parallel().mapToDouble(i -> (target[i] * Math.log(out[i] + 1e-16))).sum();
 
 		return -cost;
 	}
 
-	public double[][] derivative(double[][] output, double[][] target, Activation activation) {
-		double[][] delta = new double[output.length][output[0].length];
+	public float[][] derivative(float[][] output, float[][] target, Activation activation) {
+		float[][] delta = new float[output.length][output[0].length];
 
-		double[][] derivative = activation.derivative(output);
+		float[][] derivative = activation.derivative(output);
 
 		IntStream.range(0, output.length).parallel().forEach(b -> {
 			for (int i = 0; i < output[0].length; i++) {
