+++ src/main/neuralnet/costs/MeanSquareError.java
@@ -10,19 +10,19 @@ public class MeanSquareError implements Cost {
 		return CostType.CROSS_ENTROPY;
 	}
 
-	public double cost(double[] out, double[] target) {
-		double cost = IntStream.range(0, target.length).parallel().mapToDouble(i -> Math.pow(target[i] - out[i], 2)).sum();
+	public float cost(float[] out, float[] target) {
+		float cost = (float) IntStream.range(0, target.length).parallel().mapToDouble(i -> Math.pow(target[i] - out[i], 2)).sum();
 
-		return 0.5 * cost;
+		return 0.5f * cost;
 	}
 
-	public double[][] derivative(double[][] output, double[][] target, Activation activation) {
-		double[][] delta = new double[output.length][output[0].length];
+	public float[][] derivative(float[][] output, float[][] target, Activation activation) {
+		float[][] delta = new float[output.length][output[0].length];
 
 		if (activation.getType() == ActivationType.SOFTMAX)
 			throw new UnsupportedOperationException();
 
-		double[][] derivative = activation.derivative(output);
+		float[][] derivative = activation.derivative(output);
 
 		IntStream.range(0, output.length).parallel().forEach(b -> {
 			for (int i = 0; i < output[0].length; i++) {
