+++ src/test/neuralnet/layers/GRUTest.java
@@ -13,7 +13,8 @@ class GRUTest {
 		// just a regular test
 		Model model = new Model.Builder()
 				.add(new GRU.Builder().hiddenSize(5).build()).inputDimensions(2).cost(CostType.MEAN_SQUARE_ERROR).build();
-		assertTrue(model.gradientCheck(new double[][]{{0.3, 0.7}, {0.5, 0.2}}, new double[][]{{0.2, 0.3, 0.6, 0.1, 0.8}, {0.8, 0.7, 0.5,
-				0.6, 0.2}}));
+		assertTrue(model.gradientCheck(new float[][]{{0.3f, 0.7f}, {0.5f, 0.2f}}, new float[][]{{0.2f, 0.3f, 0.6f, 0.1f, 0.8f}, {0.8f,
+			0.7f, 0.5f,
+			0.6f, 0.2f}}));
 	}
 }
